[web] Renamed domain to src
