+++ packages/web/pages/sessions/[code].tsx
@@ -1,7 +1,5 @@
 import axios from "axios";
 import { NextPage } from "next";
-import { useRouter } from "next/router";
-import { Session } from "../../domain/session/types";
 
 interface Props {
   code?: string | string[];
@@ -9,10 +7,8 @@ interface Props {
 
 const SessionPage: NextPage<Props> = ({ code }) => <main>Your code: {code}</main>;
 
-SessionPage.getInitialProps = async ({ query, req }) => {
+SessionPage.getInitialProps = async ({ query }) => {
   const code = typeof query.code === "string" ? query.code : query.code.join("");
-
-  const userAgent = req ? req.headers["user-agent"] : navigator.userAgent;
 
   const sessionContent = await axios.get(
     `${process.env.NEXT_PUBLIC_API_HOST}/sessions/${encodeURIComponent(code)}/code`
