+++ packages/web/src/reducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers, Reducer } from "redux";
-import { sessionReducer } from "./domain/session/reducer";
-import { SessionState } from "./domain/session/types";
+import { sessionReducer } from "./session/reducer";
+import { SessionState } from "./session/types";
 
 export const reducer: Reducer<State> = combineReducers({
   session: sessionReducer
