+++ packages/web/src/saga.ts
@@ -1,6 +1,6 @@
 import createSagaMiddleware from "redux-saga";
 import { all, fork } from "redux-saga/effects";
-import { saga as sessionSaga } from "./domain/session/sagas";
+import { saga as sessionSaga } from "./session/sagas";
 
 export const sagaMiddleware = createSagaMiddleware();
